Rename date_updated to date_posted_or_crawled
The date is now only set when the post is crawled for the first time.
Updates happen silently, because they are currently not used anywhere.

This also chooses the minimum of datetime.now() and the post's date when
setting the date for a Post.  This ensures that future dated posts
aren't always marked as "latest".

Fixes #114