Ignore post published and/or updated dates
Sort posts by created_at (timestamp of creating post in db), instead. The actual
timestamp of when the post was made isn't being used anywhere and sometimes is
not set by authors. So, we could just do away with all the hackery around it, to
put in some value into the DB.