Upgrade from Python2 to Python3
Used the 2to3 converter to convert code automatically, and then made a bunch
of changes by hand.

Some of the requirements have been updated to work with Python3.

Remove gevent monkey patching since it seems to fail in Python3, this needs
to be investigated further.

A new migration file was generated during this conversion. Not sure what
exactly caused this to be required, though.