deps: Upgrade django-storages to the latest version
Upgrade to a version that works with the current version of Django. The
previous version doesn't work with the updated version of Django, but it
doesn't cause an issue with running the code locally, since the S3 storage is
only used in staging or production.

Running `python manage.py collectstatic --noinput` fails when the S3 backend is
set. This is run on Heroku deployments. It may be worth adding a test locally
to detect this, without needing to deploy.